refactored Analyser useEffect
